ISAPP-74 Package restructuring
