+++ src/main/java/vsu/cs/is/infsysserver/employee/adapter/rest/EmployeeController.java
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import vsu.cs.is.infsysserver.employee.EmployeeServiceImplementation;
+import vsu.cs.is.infsysserver.employee.adapter.rest.api.EmployeeApi;
 import vsu.cs.is.infsysserver.employee.adapter.rest.dto.request.EmployeeCreateRequest;
 import vsu.cs.is.infsysserver.employee.adapter.rest.dto.request.EmployeeUpdateRequest;
 import vsu.cs.is.infsysserver.employee.adapter.rest.dto.response.EmployeeAdminResponse;
