--- src/main/java/vsu/cs/is/infsysserver/security/entity/temp/User.java
@@ -1,56 +1,0 @@
-package vsu.cs.is.infsysserver.security.entity.temp;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.SequenceGenerator;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import vsu.cs.is.infsysserver.security.entity.token.Token;
-
-import java.util.List;
-
-@Entity
-@Data
-@Table(name = "users")
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class User {
-
-    @Id
-    @GeneratedValue(
-            strategy = GenerationType.SEQUENCE,
-            generator = "users_sequence")
-    @SequenceGenerator(
-            name = "users_sequence",
-            allocationSize = 1)
-    @Column(name = "id")
-    private Long id;
-
-    @Column(name = "first_name")
-    private String firstName;
-
-    @Column(name = "last_name")
-    private String lastName;
-
-    @Column(name = "password")
-    private String password;
-
-    @Column(name = "email")
-    private String email;
-
-    @Enumerated(EnumType.STRING)
-    private Role role;
-
-    @OneToMany(mappedBy = "user")
-    private List<Token> tokens;
-}
