--- src/main/java/vsu/cs/is/infsysserver/security/entity/temp/UserRepository.java
@@ -1,11 +1,0 @@
-package vsu.cs.is.infsysserver.security.entity.temp;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
