+++ src/main/java/vsu/cs/is/infsysserver/security/service/UserDetailsServiceImpl.java
@@ -5,7 +5,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import vsu.cs.is.infsysserver.security.entity.temp.UserRepository;
+import vsu.cs.is.infsysserver.user.adapter.jpa.UserRepository;
 import vsu.cs.is.infsysserver.security.util.UserMapper;
 
 import java.util.NoSuchElementException;
