+++ src/main/java/vsu/cs/is/infsysserver/security/util/UserMapper.java
@@ -2,7 +2,7 @@ package vsu.cs.is.infsysserver.security.util;
 
 import org.springframework.security.core.userdetails.UserDetails;
 import vsu.cs.is.infsysserver.security.entity.UserDetailsImpl;
-import vsu.cs.is.infsysserver.security.entity.temp.User;
+import vsu.cs.is.infsysserver.user.adapter.jpa.entity.User;
 
 public final class UserMapper {
