+++ src/main/java/vsu/cs/is/infsysserver/staticpage/adapter/jpa/StaticPageRepository.java
@@ -1,0 +1,7 @@
+package vsu.cs.is.infsysserver.staticpage.adapter.jpa;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import vsu.cs.is.infsysserver.staticpage.adapter.jpa.entity.StaticPage;
+
+public interface StaticPageRepository extends JpaRepository<StaticPage, Long> {
+}
