+++ src/main/java/vsu/cs/is/infsysserver/user/adapter/jpa/UserRepository.java
@@ -1,0 +1,12 @@
+package vsu.cs.is.infsysserver.user.adapter.jpa;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import vsu.cs.is.infsysserver.user.adapter.jpa.entity.User;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+}
